@@ -5,63 +5,73 @@ import org.springframework.data.mongodb.core.mapping.Document;
 
 @Document(collection = "Admin")
 public class Admin {
-	private String	nome;
-	private String	cognome;
-	private	String	username;
+    private String nome;
+    private String cognome;
+    private String username;
 
-	//MODIFICA (15/02/2024) : Aggiunta campo email
-	@Id
-	private String email;
+    //MODIFICA (15/02/2024) : Aggiunta campo email
+    @Id
+    private String email;
 
-	//MODIFICA (15/02/2024) : Aggiunta campo resetToken
-	private String resetToken;
+    //MODIFICA (15/02/2024) : Aggiunta campo resetToken
+    private String resetToken;
 
-	//MODIFICA (16/02/2024) : Aggiunta campo invitationToken
-	private String invitationToken;
+    //MODIFICA (16/02/2024) : Aggiunta campo invitationToken
+    private String invitationToken;
 
-	//FINE MODIFICA (15/02/2024)
-	private String password;
+    //FINE MODIFICA (15/02/2024)
+    private String password;
 
-	
-	public Admin(String nome, String cognome, String username, String email,String password) {
-		this.nome = nome;
-		this.cognome = cognome;
-		this.username = username;
-		this.email = email;
-		this.password = password;
-	}
-	 
-	public String getNome() {
-		return nome;
-	}
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-	public String getCognome() {
-		return cognome;
-	}
-	public void setCognome(String cognome) {
-		this.cognome = cognome;
-	}
-	public String getUsername() {
-		return username;
-	}
-	public void setUsername(String username) {
-		this.username = username;
-	}
-	public String getPassword() {
-		return password;
-	}
-	public void setPassword(String password) {
-		this.password = password;
-	}
-	public String getEmail() {
-		return email;
-	}
-	public void setEmail(String email) {
-		this.email = email;
-	}
-	// Getter e setter per il campo resetToken
+
+    public Admin(String nome, String cognome, String username, String email, String password) {
+        this.nome = nome;
+        this.cognome = cognome;
+        this.username = username;
+        this.email = email;
+        this.password = password;
+    }
+
+    public String getNome() {
+        return nome;
+    }
+
+    public void setNome(String nome) {
+        this.nome = nome;
+    }
+
+    public String getCognome() {
+        return cognome;
+    }
+
+    public void setCognome(String cognome) {
+        this.cognome = cognome;
+    }
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    // Getter e setter per il campo resetToken
     public String getResetToken() {
         return resetToken;
     }
@@ -69,8 +79,9 @@ public class Admin {
     public void setResetToken(String resetToken) {
         this.resetToken = resetToken;
     }
-	//Getter e setter per il campo invitationToken
-	public String getInvitationToken() {
+
+    //Getter e setter per il campo invitationToken
+    public String getInvitationToken() {
         return invitationToken;
     }
 
@@ -78,17 +89,17 @@ public class Admin {
         this.invitationToken = invitationToken;
     }
 
-	
-	@Override
+
+    @Override
     public String toString() {
         return "Admin{" +
                 "nome='" + nome + '\'' +
                 ", cognome='" + cognome + '\'' +
                 ", username='" + username + '\'' +
-				", email='" + email + '\'' +
+                ", email='" + email + '\'' +
                 ", password='" + password + '\'' +
-				", resetToken='" + resetToken + '\'' +
-				", invitationToken='" + invitationToken + '\'' +
+                ", resetToken='" + resetToken + '\'' +
+                ", invitationToken='" + invitationToken + '\'' +
                 '}';
     }
 }
