@@ -1,9 +1,8 @@
 package com.groom.manvsclass.model.repository;
 
+import com.groom.manvsclass.model.interaction;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-import com.groom.manvsclass.model.interaction;
-
-public interface InteractionRepository	extends MongoRepository<interaction,String>{
+public interface InteractionRepository extends MongoRepository<interaction, String> {
 
 }
