@@ -1,12 +1,10 @@
 package com.groom.manvsclass.service;
 
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-import java.util.stream.Collectors;
-
+import com.groom.manvsclass.api.ApiGatewayClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.*;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
@@ -14,38 +12,42 @@ import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestClientException;
 import org.springframework.web.client.RestTemplate;
 
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
+import java.util.stream.Collectors;
+
 @Service
 public class NotificationService {
 
     private static final Logger logger = LoggerFactory.getLogger(NotificationService.class);
     private final RestTemplate restTemplate;
+    private final ApiGatewayClient apiGatewayClient;
+
+    public NotificationService(RestTemplate restTemplate, ApiGatewayClient apiGatewayClient) {
+        this.restTemplate = restTemplate;
+        this.apiGatewayClient = apiGatewayClient;
+    }
 
     // Iniezione di RestTemplate tramite costruttore
+    /*
     public NotificationService(RestTemplate restTemplate) {
         this.restTemplate = restTemplate;
     }
 
-    public String sendNotification(String email, Integer studentID, String title, String message, String type) {
-        String url = "http://t23-g1-app-1:8080/new_notification";
+     */
 
+    public String sendNotification(String email, Integer studentID, String title, String message, String type) {
         // Verifica che almeno uno dei due identificatori sia fornito
         if (email == null && studentID == null) {
             logger.warn("Tentativo di invio notifica senza email né studentID.");
             return "Errore: Devi fornire almeno un identificatore (email o studentID)";
         }
 
-        // Headers della richiesta
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_FORM_URLENCODED);
-
         // Creazione dei parametri della richiesta
         MultiValueMap<String, String> params = prepareParams(email, studentID, title, message, type);
 
-        // Crea la richiesta HTTP
-        HttpEntity<MultiValueMap<String, String>> entity = new HttpEntity<>(params, headers);
-
         try {
-            ResponseEntity<String> response = restTemplate.exchange(url, HttpMethod.POST, entity, String.class);
+            ResponseEntity<String> response = apiGatewayClient.callSendNotification(params);
 
             if (response.getStatusCode() == HttpStatus.OK) {
                 return "Notifica inviata con successo!";
