@@ -1,13 +1,13 @@
 /*
  *   Copyright (c) 2024 Stefano Marano https://github.com/StefanoMarano80017
  *   All rights reserved.
-
+ *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
  *   You may obtain a copy of the License at
-
- *   http://www.apache.org/licenses/LICENSE-2.0
-
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
  *   Unless required by applicable law or agreed to in writing, software
  *   distributed under the License is distributed on an "AS IS" BASIS,
  *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,38 +15,40 @@
  *   limitations under the License.
  */
 
-/*
-* 		Qui ci sono le chiamate REST dell'editor 
+/* REST_Editor.js
+   Gestione delle chiamate REST per l’editor.
+   (Dipende da Util_Editor.js, che deve essere caricato prima.)
 */
-// Funzione per ottenere i dati dal localStorage
-function getGameData() {
-    let underTestClassName = localStorage.getItem("underTestClassName");
-    if (underTestClassName) {
-        // La stringa non è né nulla né vuota
-        let ClassUT = getParameterByName("ClassUT");
-        if(ClassUT === underTestClassName){
-            return {
-                playerId: 			String(parseJwt(getCookie("jwt")).userId),
-                type_robot: 		localStorage.getItem("robot"),
-                difficulty: 		localStorage.getItem("difficulty"),
-                mode: 				localStorage.getItem("modalita"),
-                underTestClassName: localStorage.getItem("underTestClassName"),
-            };
-        }
-    }   
-    //modal che ti blocca
-    openModalError(
-        "Accesso Illegale all'editor ",
-        `Non sei passato da Gamemode per settare la tua partita o hai inserito un URL sbagliato.`,
-        [{ text: 'Vai alla Home', href: '/main', class: 'btn btn-primary' }]
-    );
+
+async function getGameActionRequestBody() {
+    console.log("getGameActionRequestBody");
+    console.log(editor_robot.getValue());
+    let requestBody = {
+        playerId: jwtData.userId,
+        mode: GetMode(),
+        testingClassCode: editor_utente.getValue(),
+        classUTCode: editor_robot.getValue()
+    };
+
+    if (GetMode() === "PartitaSingola")
+        requestBody["remainingTime"] = timer_remainingTime.toString();
+
+	return requestBody;
 }
 
-// Funzione per eseguire la richiesta AJAX
-async function runGameAction(url, formData, isGameEnd) {
+// Funzione per salvare la sessione quando l'utente lascia la pagina (necessaria per implementare il timer)
+async function handleBeforeUnload(e) {
+    let leaveGameRequestBody = await getGameActionRequestBody();
+    console.log("[POST /leave] Sending request upon leaving");
+    navigator.sendBeacon("/api/gameEngine/leave", JSON.stringify(leaveGameRequestBody));
+}
+window.addEventListener("beforeunload", handleBeforeUnload);
+
+// === FUNZIONE PER LA RICHIESTA AJAX DEL GIOCO ===
+async function runGameAction(url, runGameRequestBody) {
     try {
-        formData.append("isGameEnd", isGameEnd);
-        const response = await ajaxRequest(url, "POST", formData, false, "json");
+        console.log("runGameRequestBody", runGameRequestBody);
+        const response = await ajaxRequest_ForRun(url, "POST", runGameRequestBody, false, "json");
         return response;
     } catch (error) {
         console.error("Errore nella richiesta AJAX:", error);
@@ -54,105 +56,179 @@ async function runGameAction(url, formData, isGameEnd) {
     }
 }
 
-// Documento pronto
 $(document).ready(function () {
-    const data = getGameData();
-    startGame(data);
-    if (data.mode === "Allenamento") {
-        document.getElementById("runButton").disabled = true;
-        console.log("Sei in allenamento");
-    }
     const savedContent = localStorage.getItem('codeMirrorContent');
-	if (!savedContent) {
-        //non ho salvato il contenuto dell'editor, quindi lo scarico e imposto il template con i dati dell'utente
-        let formattedDate = `${String(currentDate.getDate()).padStart(2, "0")}/${String(currentDate.getMonth() + 1).padStart(2, "0")}/${currentDate.getFullYear()}`;
-        let replacements  = {
-            TestClasse: `Test${localStorage.getItem("underTestClassName")}`,
+    if (!savedContent) {
+        const currentDate = new Date();
+        const formattedDate = `${String(currentDate.getDate()).padStart(2, "0")}/${String(currentDate.getMonth() + 1).padStart(2, "0")}/${currentDate.getFullYear()}`;
+        const replacements = {
+            TestClasse: `Test${ClassName}`,
             username: jwtData.sub,
             userID: jwtData.userId,
             date: formattedDate,
         };
         SetInitialEditor(replacements);
-	}else{
-        //Se ho del contenuto salvato 
+    } else {
         editor_utente.setValue(savedContent);
-        //document.getElementById('Editor_utente').value = savedContent;
-        editor_utente.refresh(); // Ricarica l'editor per applicare le modifiche
+        editor_utente.refresh();
     }
-    const savedStorico = localStorage.getItem('storico');
-    if(savedStorico){
+    
+    if (localStorage.getItem('storico')) {
         viewStorico();
     }
 });
 
-
 let isActionInProgress = false; // Flag per indicare se un'azione è attualmente in corso
 
 // Funzione principale per gestire l'azione del gioco
-async function handleGameAction(isGameEnd) {
+async function handleGameAction(isGameEnd, compileUponEndTime=false) {
     isActionInProgress = true; // Imposta il flag per bloccare altre azioni
     run_button.disabled = true; // Disabilita il pulsante di esecuzione
     coverage_button.disabled = true; // Disabilita il pulsante di coverage
-
     // Determina le chiavi per il caricamento e il pulsante in base a isGameEnd
     const loadingKey = isGameEnd ? "loading_run" : "loading_cov";
     const buttonKey = isGameEnd ? "runButton" : "coverageButton";
+    // Mostra l'indicatore di caricamento
+    toggleLoading(true, loadingKey, buttonKey);
+    // Aggiorna lo stato a "sending"
+    setStatus("sending");
+    // Otteniamo il requestBody (con debug sul codice dell'editor)
+    let requestBody = await getGameActionRequestBody()
 
-    toggleLoading(true, loadingKey, buttonKey); // Mostra l'indicatore di caricamento
-    setStatus("sending"); // Aggiorna lo stato a "sending"
+    if (isGameEnd) {
+        try {
+            //Esegue la terminazione del gioco
 
-    const formData = getFormData(); // Recupera i dati del modulo
-    const response = await runGameAction("/run", formData, isGameEnd); // Esegue l'azione del gioco
-    setStatus("compiling"); // Aggiorna lo stato a "compiling"
+            if (!compileUponEndTime) {
+                requestBody["testingClassCode"] = "";
+                setStatus("game_end");
+                const endResponse = await runGameAction("/api/gameEngine/EndGame", requestBody);
 
-    handleResponse(response, formData, isGameEnd, loadingKey, buttonKey); // Gestisce la risposta ricevuta
+                console.error(endResponse);
 
-    isActionInProgress = false; // Reimposta il flag al termine dell'azione
+                handleGameRun(endResponse.runGameResponse, loadingKey, buttonKey, true);
+                handleGameEnd(endResponse);
+                toggleLoading(false, loadingKey, buttonKey);
+            } else {
+                const endResponse = await runGameAction("/api/gameEngine/EndGame", requestBody);
+
+                handleGameRun(endResponse.runGameResponse, loadingKey, buttonKey, true);
+                handleGameEnd(endResponse);
+                toggleLoading(false, loadingKey, buttonKey);
+            }
+        } catch (error) {
+            console.error("[handleGameAction] Errore durante l'esecuzione:", error);
+            handleServerInternalError(error, loadingKey, buttonKey);
+        }
+    } else {
+        try {
+            //Esegue l'azione di gioco
+            const response = await runGameAction("/api/gameEngine/run", requestBody);
+            setStatus("compiling");
+            handleGameRun(response, loadingKey, buttonKey, false);
+            resetButtons();
+        } catch (error) {
+            handleServerInternalError(error, loadingKey, buttonKey);
+        }
+    }
+    isActionInProgress = false;
 }
 
-// Gestisce la risposta dal server
-function handleResponse(response, formData, isGameEnd, loadingKey, buttonKey) {
-    const { robotScore, userScore, outCompile, 
-            coverage, gameId, roundId,
-            coverageDetails, isWinner} = response;
-    // Aggiorna i dati del modulo con gameId e roundId
-    formData.append("gameId", gameId);
-    formData.append("roundId", roundId);
-    console_utente.setValue(outCompile); // Mostra l'output della compilazione nella console utente
-    parseMavenOutput(outCompile); // Analizza l'output di Maven
-    if (!coverage) { // Se non c'è copertura, gestisce l'errore di compilazione
+function handleServerInternalError(error, loadingKey, buttonKey) {
+    console.error("[handleGameAction] Errore durante l'esecuzione:", error);
+    const errorObject = JSON.parse(error.responseText);
+
+    try {
+        switch (error.status) {
+            case 429:
+                console_robot.setValue(errorMessages["429"]);
+                break;
+            case 504:
+                console_robot.setValue(errorMessages["504"]);
+                break;
+            default:
+                const formattedError = JSON.stringify(errorObject, null, 4);
+                console_robot.setValue(formattedError);
+                break;
+        }
+    } catch (e) {
+        // Se non è un JSON valido stampo l'errore raw
+        console_robot.setValue(error.responseText);
+    }
+
+    toggleLoading(false, loadingKey, buttonKey); // Nasconde l'indicatore di caricamento
+    resetButtons(); // Reimposta i pulsanti
+}
+
+
+function handleGameEnd(response) {
+    let {userScore, robotScore, isWinner, expGained, achievementsUnlocked} = response;
+    console.log("achievementsUnlocked", achievementsUnlocked);
+
+    generateEndGameMessage(userScore, robotScore, isWinner, expGained, achievementsUnlocked); // Gestisce la fine del gioco
+
+    // Disattivo il timer
+    if (GetMode() === "PartitaSingola")
+        stopTimer();
+
+    // Disattivo la chiamata a POST /leave all'uscita dalla pagina
+    window.removeEventListener("beforeunload", handleBeforeUnload);
+}
+
+function handleGameRun(response, loadingKey, buttonKey, isGameEnd) {
+    const {
+            userCoverageDetails, robotCoverageDetails,
+            canWin, unlockedAchievements,
+            userScore, robotScore,
+        } = response;
+
+    console.log("game run", response);
+
+    const userOutputCompile = userCoverageDetails.compileOutput;
+    const userCoverage_ForHighlight = userCoverageDetails.xml_coverage;
+    const robotCoverage_ForHighlight = robotCoverageDetails.xml_coverage;
+
+    console.log("userOutputCompile", userOutputCompile);
+    console.log("userCoverage_ForHighlight", userCoverageDetails.xml_coverage);
+    console.log("robotCoverage_ForHighlight", robotCoverageDetails.xml_coverage);
+
+    console_utente.setValue(userOutputCompile); // Mostra l'output della compilazione nella console utente
+    parseMavenOutput(userOutputCompile); // Analizza l'output di Maven
+    if (!userCoverage_ForHighlight) { // Se non c'è copertura, gestisce l'errore di compilazione
         setStatus("error");
-        handleCompileError(loadingKey, buttonKey); // Gestisce l'errore
+        //Gestione degli errori 
+        handleCompileError(loadingKey, buttonKey);
         return;
     }
+
     // Se la copertura è disponibile, la processa
-    processCoverage(coverage, formData, robotScore, userScore, isGameEnd, loadingKey, buttonKey, coverageDetails, isWinner);
+    processCoverage(
+        userCoverage_ForHighlight, robotCoverage_ForHighlight,
+        userCoverageDetails, robotCoverageDetails,
+        userScore, robotScore,
+        canWin, unlockedAchievements,
+        loadingKey, buttonKey, isGameEnd
+    );
 }
 
 // Processa la copertura del codice e aggiorna i dati di gioco
-async function processCoverage(coverage, formData, robotScore, userScore, isGameEnd, loadingKey, buttonKey, coverageDetails, isWinner) {
-    highlightCodeCoverage($.parseXML(coverage), editor_robot); // Evidenzia la copertura del codice nell'editor
+async function processCoverage(userCoverage_ForHighlight, robotCoverage_ForHighlight, userCoverageDetails, robotCoverageDetails, userScore, robotScore, canWin, unlockedAchievements, loadingKey, buttonKey, isGameEnd) {
+    highlightCodeCoverage($.parseXML(userCoverage_ForHighlight), $.parseXML(robotCoverage_ForHighlight), editor_robot); // Evidenzia la copertura del codice nell'editor
     orderTurno++; // Incrementa l'ordine del turno
-    const csvContent = await fetchCoverageReport(formData); // Recupera il report di coverage
+
     setStatus("loading"); // Aggiorna lo stato a "loading"
-    const valori_csv = extractThirdColumn(csvContent); // Estrae i valori dalla terza colonna del CSV
-    updateStorico(orderTurno, userScore, valori_csv[0]); // Aggiorna lo storico del gioco
-    setStatus(isGameEnd ? "game_end" : "turn_end"); // Imposta lo stato di fine gioco o fine turno
+    updateStorico(orderTurno, userScore, userCoverageDetails.evosuite_line); // Aggiorna lo storico del gioco
+
+    setStatus("turn_end"); // Imposta lo stato di fine gioco o fine turno
     toggleLoading(false, loadingKey, buttonKey); // Nasconde l'indicatore di caricamento
-    displayUserPoints(isGameEnd, valori_csv, robotScore, userScore, coverageDetails, iswinner); // Mostra i punti dell'utente
-    if (isGameEnd) { // Se il gioco è finito
-        handleEndGame(userScore); // Gestisce la fine del gioco
-    } else {
-        resetButtons(); // Reimposta i pulsanti
-    }
+    displayUserPoints(userCoverageDetails, robotCoverageDetails, canWin, userScore, robotScore, isGameEnd); // Mostra i punti dell'utente
+    if (!isGameEnd && unlockedAchievements.length !== 0)
+        handleUnlockedAchievements(unlockedAchievements);
 }
 
 // Mostra i punti dell'utente nella console
-function displayUserPoints(isGameEnd, valori_csv, robotScore, userScore, coverageDetails, isWinner) {
-    const displayUserPoints = isGameEnd 
-        ? getConsoleTextRun(valori_csv, coverageDetails, robotScore, userScore, isWinner) // Testo per la fine del gioco
-        : getConsoleTextCoverage(valori_csv, userScore, coverageDetails); // Testo per la copertura
-
+function displayUserPoints(userCoverageDetails, robotCoverageDetails, canWin, userScore, robotScore, isGameEnd) {
+    const displayUserPoints = getConsoleTextRun(userCoverageDetails, robotCoverageDetails, canWin, userScore, robotScore, isGameEnd)
     console_robot.setValue(displayUserPoints); // Aggiorna la console del robot con i punti
 }
 
@@ -163,25 +239,49 @@ function handleCompileError(loadingKey, buttonKey) {
     resetButtons(); // Reimposta i pulsanti
 }
 
-// Recupera il report di coverage da T8
-async function fetchCoverageReport(formData) {
-    const url = createApiUrl(formData, orderTurno); // Crea l'URL dell'API
-    return await ajaxRequest(url, "POST", formData.get("testingClassCode"), false, "text"); // Esegue la richiesta AJAX
+// Gestisce la fine del gioco, mostra un messaggio e pulisce i dati
+function generateEndGameMessage(userScore, robotScore, isWinner, expGained, achievementsUnlocked) {
+    let resultMessage = isWinner ? gameEndData.game_win : gameEndData.game_lose;
+    let expMessage = "";
+    let achievementsMessage = ""
+    let rememberSaveMessage = gameEndData.game_remember_save;
+
+    if (isWinner) {
+        if (expGained === 0) {
+            expMessage = gameEndData.game_exp.zero;
+        } else if (expGained === 1) {
+            expMessage = `${gameEndData.game_exp.base} ${gameEndData.game_exp.one}`;
+        } else {
+            expMessage = `${gameEndData.game_exp.base} ${expGained} ${gameEndData.game_exp.multi}`;
+        }
+
+        if (achievementsUnlocked.length > 0) {
+            achievementsMessage = "\n\n" + `${unlockedNewAchievementMessage.descr}\n${achievementsUnlocked.map(a => ` - ${achievementData[a]?.name || a}\n`).join("")}`;
+            achievementsMessage = achievementsMessage.slice(0, -1);
+        }
+    } else {
+        expMessage = gameEndData.game_retry;
+    }
+
+    openModalWithText(
+        gameEndData.game_end,
+        `${gameEndData.game_score}: ${userScore} pt.\n${resultMessage}\n${expMessage}${achievementsMessage}\n\n${rememberSaveMessage}`,
+        [{ tagName: "button", text: `${modalButtonText.close}`, data_bs_dismiss: "modal", class: 'btn btn-primary' }]
+    );
 }
 
-// Gestisce la fine del gioco, mostra un messaggio e pulisce i dati
-function handleEndGame(userScore) {
+function handleUnlockedAchievements(unlockedAchievements) {
     openModalWithText(
-        status_game_end,
-        `${score_partita_text} ${userScore} pt.`, // Mostra il punteggio dell'utente
-        [{ text: vai_home, href: '/main', class: 'btn btn-primary' }] // Pulsante per tornare alla home
+        unlockedNewAchievementMessage.title,
+        `${unlockedNewAchievementMessage.descr}\n${unlockedAchievements.map(a => ` - ${achievementData[a]?.name || a}\n`).join("")}`,
+        [{ tagName: "button", text: `${modalButtonText.close}`, data_bs_dismiss: "modal", class: 'btn btn-primary' }]
     );
-    flush_localStorage(); // Pulisce i dati salvati nel localStorage
 }
+
 
 // Reimposta i pulsanti per consentire nuove azioni
 function resetButtons() {
-    run_button.disabled = (localStorage.getItem("modalita") === "Allenamento"); // Abilita/disabilita in base alla modalità
+    run_button.disabled = (mode === "Allenamento"); // Abilita/disabilita in base alla modalità
     coverage_button.disabled = false; // Abilita il pulsante di coverage
 }
 
@@ -202,6 +302,9 @@ window.addEventListener('beforeunload', (event) => {
         // Il messaggio predefinito non può essere personalizzato, ma il modal può apparire
         return ''; // Restituisce una stringa vuota per attivare il messaggio predefinito
     }
+    // Pulisco il localstorage dal test scritto dall'utente 
+    localStorage.removeItem('codeMirrorContent');
+    localStorage.removeItem('storico');
 });
 
 // Pulsante "Run/Submit"
