@@ -1,13 +1,13 @@
 /*
  *   Copyright (c) 2024 Stefano Marano https://github.com/StefanoMarano80017
  *   All rights reserved.
-
+ *
  *   Licensed under the Apache License, Version 2.0 (the "License");
  *   you may not use this file except in compliance with the License.
  *   You may obtain a copy of the License at
-
- *   http://www.apache.org/licenses/LICENSE-2.0
-
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
  *   Unless required by applicable law or agreed to in writing, software
  *   distributed under the License is distributed on an "AS IS" BASIS,
  *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,69 +15,12 @@
  *   limitations under the License.
  */
 
-
-/*
-*		QUI CI SONO TUTTE LE FUNZIONI UTILI PER L'EDITOR 
+/* UTIL_Editor.js
+   Funzioni di utilità per l'editor.
 */
 
-// Funzione per creare l'URL dell'API
-function createApiUrl(formData, orderTurno) {
-	const className = formData.get("className");
-	const underTestClassName = formData.get("underTestClassName");
-	const playerId = formData.get("playerId");
-	// Costruisce il percorso per la classe
-	const classePath = `VolumeT8/FolderTreeEvo/${className}/${className}SourceCode/${underTestClassName}`;
-	// Ottiene il percorso del test generato
-	const testPath = generaPercorsoTest(orderTurno, formData);
-	// Costruisce l'URL dell'API
-	const apiUrl = `/api/${classePath}+${testPath}+/app+${playerId}`;
-	return apiUrl;
-}
-
-// Funzione per generare il percorso del test
-function generaPercorsoTest(orderTurno, formData) {
-	let modalita = localStorage.getItem("modalita");
-	const playerId = formData.get("playerId");
-	const gameId = formData.get("gameId");
-	const roundId = formData.get("roundId");
-	const classeLocal = formData.get("className");
-
-	modalita = (modalita === "Allenamento") ? "Sfida" : modalita;
-	
-	// Verifica la modalità e costruisce il percorso appropriato
-	if (modalita === "Scalata" || modalita === "Sfida") {
-		const scalataPart =
-			modalita === "Scalata"
-				? `/${localStorage.getItem("SelectedScalata")}${localStorage.getItem(
-						"scalataId"
-				  )}`
-				: "";
-
-		return `/VolumeT8/FolderTreeEvo/StudentLogin/Player${playerId}/${modalita}${scalataPart}/${classeLocal}/Game${gameId}/Round${roundId}/Turn${orderTurno}/TestReport`;
-	} else {
-		console.error("Errore: modalità non trovata");
-		window.location.href = "/main";
-		return null;
-	}
-}
-
-// Funzione per estrarre la terza colonna di un CSV
-function extractThirdColumn(csvContent) {
-	const rows = csvContent.split("\n"); // Divide le righe
-	const thirdColumnValues = [];
-
-	// Inizia il ciclo dalla seconda riga (indice 1)
-	rows.slice(1).forEach((row) => {
-		const cells = row.split(","); // Divide le celle
-		if (cells.length >= 3) {
-			thirdColumnValues.push(cells[2]); // Aggiunge la terza colonna
-		}
-	});
-
-	return thirdColumnValues;
-}
-
-you_win = `
+// === FUNZIONI PER GENERARE IL TESTO DI OUTPUT ===
+const you_win = `
 __     ______  _    _  __          _______ _   _ 
 \\ \\   / / __ \\| |  | | \\ \\        / /_   _| \\ | |
  \\ \\_/ / |  | | |  | |  \\ \\  /\\  / /  | | |  \\| |
@@ -86,7 +29,7 @@ __     ______  _    _  __          _______ _   _
    |_|  \\____/ \\____/      \\/  \\/   |_____|_| \\_|
 `;
 
-var you_lose = `
+const you_lose = `
 __     ______  _    _   _      ____   _____ ______ 
 \\ \\   / / __ \\| |  | | | |    / __ \\ / ____|  ____|
  \\ \\_/ / |  | | |  | | | |   | |  | | (___ | |__   
@@ -95,7 +38,7 @@ __     ______  _    _   _      ____   _____ ______
    |_|  \\____/ \\____/  |______\\____/|_____/|______|
 `;
 
-var error = `
+const error = `
 ______ _____  _____   ____   _____  
 |  ____|  __ \|  __ \ / __ \ / ____| 
 | |__  | |__) | |__) | |  | | (___   
@@ -104,125 +47,158 @@ ______ _____  _____   ____   _____
 |______|_|  \_\_|  \_\\____/|_____/  
 `;
 
-function getConsoleTextCoverage(valori_csv, gameScore, coverageDetails) {
+ +
+'════════════════════════════════════════════════════════════════════════\n\n\n' +
+ + '\n\n\n';
 
-	let lineCoveragePercentage = (coverageDetails.line.covered / (coverageDetails.line.covered + coverageDetails.line.missed)) * 100;
-	let BranchCoveragePercentage = (coverageDetails.branch.covered / (coverageDetails.branch.covered + coverageDetails.branch.missed)) * 100;
-	let instructionCoveragePercentage = (coverageDetails.instruction.covered / (coverageDetails.instruction.covered + coverageDetails.instruction.missed)) * 100;
+function getConsoleTextRun(userCoverageDetails, robotCoverageDetails, canWin, userScore, robotScore, isGameEnd) {
+	const maxLineLength = 62; // Ridotto per accomodare 4 spazi laterali
 
-	var consoleText = 
-`============================== Results ===============================
-Il tuo punteggio: ${gameScore}pt
-============================== JaCoCo ===============================
-Line Coverage COV%:  ${lineCoveragePercentage}% LOC
-covered: ${coverageDetails.line.covered}  
-missed: ${coverageDetails.line.missed}
-----------------------------------------------------------------------
-Branch Coverage COV%:  ${BranchCoveragePercentage}% LOC
-covered: ${coverageDetails.branch.covered} 
-missed: ${coverageDetails.branch.missed}
-----------------------------------------------------------------------
-Instruction Coverage COV%:  ${instructionCoveragePercentage}% LOC
-covered: ${coverageDetails.instruction.covered} 
-missed: ${coverageDetails.instruction.missed}
-============================== EvoSuite ===============================
-la tua Coverage:  ${valori_csv[0]*100}% LOC
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[1]*100}% Branch
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[2]*100}% Exception
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[3]*100}% WeakMutation
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[4]*100}% Output
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[5]*100}% Method
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[6]*100}% MethodNoException
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[7]*100}% CBranch
-======================================================================`;
+	function wrapAtSpaces(str) {
+		const limit = maxLineLength;
+		let result = '';
+		let i = 0;
 
-	// Restituisce il testo generato
-	return consoleText;
-}
+		while (i < str.length) {
+			let end = i + limit;
+			if (end >= str.length) {
+				result += '    ' + str.slice(i);
+				break;
+			}
 
-function getConsoleTextRun(valori_csv, coverageDetails, punteggioRobot, gameScore, isWinner) {
-	let lineCoveragePercentage = (coverageDetails.line.covered / (coverageDetails.line.covered + coverageDetails.line.missed)) * 100;
-	let BranchCoveragePercentage = (coverageDetails.branch.covered / (coverageDetails.branch.covered + coverageDetails.branch.missed)) * 100;
-	let instructionCoveragePercentage = (coverageDetails.instruction.covered / (coverageDetails.instruction.covered + coverageDetails.instruction.missed)) * 100;
-	
-	var consoleText2 = (isWinner) ? you_win : you_lose;
-	consoleText =
-`===================================================================== \n` +
-		consoleText2 +
-		"\n" +
-`============================== Results ===============================
-Il tuo punteggio:${gameScore}pt
-----------------------------------------------------------------------
-La coverage del robot:${punteggioRobot}% LOC
-============================== JaCoCo ===============================
-Line Coverage COV%:  ${lineCoveragePercentage}% LOC
-covered: ${coverageDetails.line.covered}  
-missed: ${coverageDetails.line.missed}
-----------------------------------------------------------------------
-Branch Coverage COV%:  ${BranchCoveragePercentage}% LOC
-covered: ${coverageDetails.branch.covered} 
-missed: ${coverageDetails.branch.missed}
-----------------------------------------------------------------------
-Instruction Coverage COV%:  ${instructionCoveragePercentage}% LOC
-covered: ${coverageDetails.instruction.covered} 
-missed: ${coverageDetails.instruction.missed}
-============================== EvoSuite ===============================
-la tua Coverage:  ${valori_csv[0]*100}% LOC
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[1]*100}% Branch
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[2]*100}% Exception
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[3]*100}% WeakMutation
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[4]*100}% Output
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[5]*100}% Method
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[6]*100}% MethodNoException
-----------------------------------------------------------------------
-Il tuo punteggio EvoSuite: ${valori_csv[7]*100}% CBranch
-======================================================================`;
+			let spaceIndex = str.lastIndexOf(' ', end);
+			if (spaceIndex <= i) {
+				spaceIndex = str.indexOf(' ', end);
+				if (spaceIndex === -1) {
+					result += '    ' + str.slice(i);
+					break;
+				}
+			}
 
-	// Restituisce il testo generato
-	return consoleText;
+			result += '    ' + str.slice(i, spaceIndex) + '\n';
+			i = spaceIndex + 1;
+		}
+
+		return result;
+	}
+
+	function roundToTwoDecimals(value) {
+		return Math.round(value * 100) / 100;
+	}
+
+	function centerText(text, width) {
+		const totalPadding = width - text.length;
+		const paddingLeft = Math.floor(totalPadding / 2);
+		const paddingRight = totalPadding - paddingLeft;
+		return ' '.repeat(paddingLeft) + text + ' '.repeat(paddingRight);
+	}
+
+	function formatJacocoRow(label, user, robot) {
+		const userPct = roundToTwoDecimals(user.covered / (user.covered + user.missed) * 100);
+		const robotPct = roundToTwoDecimals(robot.covered / (robot.covered + robot.missed) * 100);
+		const userTotal = user.covered + user.missed;
+		const robotTotal = robot.covered + robot.missed;
+
+		const userStr = `${userPct.toFixed(2)}% (${user.covered}/${userTotal} LOC)`;
+		const robotStr = `${robotPct.toFixed(2)}% (${robot.covered}/${robotTotal} LOC)`;
+
+		return `    ${label.padEnd(14)} | ${userStr.padStart(21)} | ${robotStr.padStart(21)}`;
+	}
+
+	function formatEvoRow(label, user, robot) {
+		let userPct = 0, robotPct = 0;
+
+		const robotTotal = robot.covered + robot.missed;
+		const userTotal = user.covered + user.missed;
+
+		if (robotTotal > 0) {
+			userPct = userTotal > 0 ? roundToTwoDecimals((user.covered / userTotal) * 100) : 0;
+			robotPct = roundToTwoDecimals((robot.covered / robotTotal) * 100);
+		}
+
+		const userStr = `${userPct.toFixed(2)}%`;
+		const robotStr = `${robotPct.toFixed(2)}%`;
+
+		return `    ${label.padEnd(14)} | ${userStr.padStart(21)} | ${robotStr.padStart(21)}`;
+	}
+
+	const userInstructionCov = roundToTwoDecimals(userCoverageDetails.jacoco_instruction.covered / (userCoverageDetails.jacoco_instruction.covered + userCoverageDetails.jacoco_instruction.missed) * 100);
+	const robotInstructionCov = roundToTwoDecimals(robotCoverageDetails.jacoco_instruction.covered / (robotCoverageDetails.jacoco_instruction.covered + robotCoverageDetails.jacoco_instruction.missed) * 100);
+
+	let result =
+	wrapAtSpaces(terminalMessages.points_descr) + "\n\n\n" +
+	'    ' + '═'.repeat(maxLineLength) + '\n' +
+	`    ${terminalMessages.your_instruction_coverage}: ${userInstructionCov}% LOC\n` +
+	`    ${terminalMessages.robot_instruction_coverage}: ${robotInstructionCov}% LOC\n` +
+	'    ' + '═'.repeat(maxLineLength) + '\n\n\n' +
+	wrapAtSpaces(canWin ? terminalMessages.you_can_win : terminalMessages.you_cant_win) + '\n\n\n\n';
+
+	const header =
+		`    ${centerText(terminalMessages.metrics, 14)} |` +
+		` ${centerText(terminalMessages.you, 21)} |` +
+		` ${centerText(terminalMessages.robot, 21)}\n` +
+		`    ${'─'.repeat(14)} | ${'─'.repeat(21)} | ${'─'.repeat(21)}\n`;
+
+	// JaCoCo
+	result +=
+		'    ' + '─'.repeat(maxLineLength) + '\n' +
+		`    ${centerText(terminalMessages.jacoco_table_title, maxLineLength)}\n` +
+		'    ' + '─'.repeat(maxLineLength) + '\n' +
+		header +
+		formatJacocoRow('Instruction', userCoverageDetails.jacoco_instruction, robotCoverageDetails.jacoco_instruction) + '\n' +
+		formatJacocoRow('Line', userCoverageDetails.jacoco_line, robotCoverageDetails.jacoco_line) + '\n' +
+		formatJacocoRow('Branch', userCoverageDetails.jacoco_branch, robotCoverageDetails.jacoco_branch) + '\n\n\n\n';
+
+	// EvoSuite
+	console.log("isGameEnd", isGameEnd)
+	if (isGameEnd) {
+		result +=
+			'    ' + '─'.repeat(maxLineLength) + '\n' +
+			`    ${centerText(terminalMessages.evosuite_table_title, maxLineLength)}\n` +
+			'    ' + '─'.repeat(maxLineLength) + '\n' +
+			header +
+			formatEvoRow('Line', userCoverageDetails.evosuite_line, robotCoverageDetails.evosuite_line) + '\n' +
+			formatEvoRow('Branch', userCoverageDetails.evosuite_branch, robotCoverageDetails.evosuite_branch) + '\n' +
+			formatEvoRow('Exception', userCoverageDetails.evosuite_exception, robotCoverageDetails.evosuite_exception) + '\n' +
+			formatEvoRow('WeakMutation', userCoverageDetails.evosuite_weak_mutation, robotCoverageDetails.evosuite_weak_mutation) + '\n' +
+			formatEvoRow('CBranch', userCoverageDetails.evosuite_cbranch, robotCoverageDetails.evosuite_cbranch) + '\n\n';
+	}
+
+	return result;
 }
 
 function getConsoleTextError(){
 	return  `===================================================================== \n` 
 			+ error +  "\n" +
-			`============================== Results =============================== \n
-			Ci sono stati errori di compilazione, controlla la console !`;
+			`============================== ${terminalMessages.results} =============================== \n
+			${terminalMessages.compilation_error}`;
 }
 
-// Funzione per avviare il gioco utilizzando ajaxRequest
-async function startGame(data) {
-	try {
-		// Utilizziamo la funzione ajaxRequest per la chiamata POST
-		const response = await ajaxRequest(
-			"/StartGame",
-			"POST",
-			data,
-			true,
-			"text"
-		);
-		console.log("Partita iniziata con successo:", response);
-	} catch (error) {
-		// Assicurati che error sia una stringa prima di usare includes
-		errorMessage = JSON.stringify(error);
-		if (errorMessage.includes("errore esiste già la partita")) {
-			console.log("Il messaggio d'errore indica che esiste già la partita.");
-			openModalWithText(status_inizia_partita, match_in_corso);
-		} 
-		console.log("[start Game]", errorMessage);
-	}
+// Funzione per analizzare l'output di Maven
+function parseMavenOutput(output) {
+    const lines = output.split('\n');
+    let results = {
+        errors: 0,
+        warnings: 0
+    };
+
+    lines.forEach(line => {
+        // Regex per contare avvisi
+        const warningMatch = line.match(/^\[INFO\] (\d+) warning/);
+        if (warningMatch) {
+            results.warnings = parseInt(warningMatch[1], 10);
+        }
+
+        // Regex per contare errori
+        const errorMatch = line.match(/^\[INFO\] (\d+) error/);
+        if (errorMatch) {
+            results.errors = parseInt(errorMatch[1], 10);
+        }
+    });
+
+	document.getElementById("error_compiler").textContent = results.errors;
+	document.getElementById("warning_compiler").textContent =  results.warnings;
+    return results;
 }
 
 function toggleLoading(showSpinner, divId, buttonId) {
@@ -285,66 +261,127 @@ function setStatus(statusName) {
     }
 }
 
-function highlightCodeCoverage(reportContent, editor) {
+function highlightCodeCoverage(reportContent, robotContent, editor) {
 	// Analizza il contenuto del file di output di JaCoCo per individuare le righe coperte, non coperte e parzialmente coperte
 	// Applica lo stile appropriato alle righe del tuo editor
 	var coveredLines = [];
 	var uncoveredLines = [];
 	var partiallyCoveredLines = [];
 
+	var coveredLinesRobot = [];
+	var uncoveredLinesRobot = [];
+	var partiallyCoveredLinesRobot = [];
+
 	reportContent.querySelectorAll("line").forEach(function (line) {
-		if (line.getAttribute("mi") == 0)
+		const mi = parseInt(line.getAttribute("mi"));
+		const ci = parseInt(line.getAttribute("ci"));
+		const mb = parseInt(line.getAttribute("mb"));
+		const cb = parseInt(line.getAttribute("cb"));
+
+		const totalInstructions = mi + ci;
+		const totalBranches = mb + cb;
+
+		const allInstructionsCovered = (mi === 0);
+		const allBranchesCovered = (totalBranches === 0 || mb === 0);
+
+		if (allInstructionsCovered && allBranchesCovered) {
 			coveredLines.push(line.getAttribute("nr"));
-		else if (line.getAttribute("mb") > 0 && line.getAttribute("cb") > 0)
+		} else if ((ci > 0 || cb > 0)) {
 			partiallyCoveredLines.push(line.getAttribute("nr"));
-		else uncoveredLines.push(line.getAttribute("nr"));
+		} else {
+			uncoveredLines.push(line.getAttribute("nr"));
+		}
+	});
+
+
+	robotContent.querySelectorAll("line").forEach(function (line) {
+		const mi = parseInt(line.getAttribute("mi"));
+		const ci = parseInt(line.getAttribute("ci"));
+		const mb = parseInt(line.getAttribute("mb"));
+		const cb = parseInt(line.getAttribute("cb"));
+
+		const totalInstructions = mi + ci;
+		const totalBranches = mb + cb;
+
+		const allInstructionsCovered = (mi === 0);
+		const allBranchesCovered = (totalBranches === 0 || mb === 0);
+
+		if (allInstructionsCovered && allBranchesCovered) {
+			coveredLinesRobot.push(line.getAttribute("nr"));
+		} else if ((ci > 0 || cb > 0)) {
+			uncoveredLinesRobot.push(line.getAttribute("nr"));
+		} else {
+			partiallyCoveredLinesRobot.push(line.getAttribute("nr"));
+		}
+	});
+
+	let decreaseRobotLines = 1;
+	const packages = robotContent.querySelectorAll("package");
+
+	for (const line of packages) {
+		if (line.getAttribute("name") !== "") {
+			decreaseRobotLines = 2;
+			break; 
+		} else if (line.getAttribute("name") === "Calcolatrice") {
+			decreaseRobotLines = 2;
+			break;
+		}
+	}
+
+	coveredLinesRobot.forEach(function (lineNumber) {
+		editor.removeLineClass(lineNumber - decreaseRobotLines, "gutter", "bg-danger");
+		editor.removeLineClass(lineNumber - decreaseRobotLines, "gutter", "bg-warning");
+		editor.addLineClass	(lineNumber - decreaseRobotLines, "gutter", "  bg-success");
+	});
+
+	uncoveredLinesRobot.forEach(function (lineNumber) {
+		editor.removeLineClass(lineNumber - decreaseRobotLines, "gutter", "bg-warning");
+		editor.removeLineClass(lineNumber - decreaseRobotLines, "gutter", "bg-success");
+		editor.addLineClass	(lineNumber - decreaseRobotLines, "gutter", "bg-danger");
+	});
+
+	partiallyCoveredLinesRobot.forEach(function (lineNumber) {
+		editor.removeLineClass(lineNumber - decreaseRobotLines, "gutter", "bg-danger");
+		editor.removeLineClass(lineNumber - decreaseRobotLines, "gutter", "bg-success");
+		editor.addLineClass	(lineNumber - decreaseRobotLines, "gutter", "bg-warning");
 	});
 
 	coveredLines.forEach(function (lineNumber) {
-		editor.removeLineClass(lineNumber - 2, "gutter", "bg-danger");
-		editor.removeLineClass(lineNumber - 2, "gutter", "bg-warning");
-		editor.addLineClass	(lineNumber - 2, "gutter", "  bg-success");
+		editor.removeLineClass(lineNumber - 3, "background", "bg-coverage-danger");
+		editor.removeLineClass(lineNumber - 3, "background", "bg-coverage-warning");
+		editor.addLineClass	(lineNumber - 3, "background", " bg-coverage-success");
 	});
 
-	uncoveredLines.forEach(function (lineNumber) { 
-		editor.removeLineClass(lineNumber - 2, "gutter", "bg-warning");
-		editor.removeLineClass(lineNumber - 2, "gutter", "bg-success");
-		editor.addLineClass	(lineNumber - 2, "gutter", "bg-danger");
+	uncoveredLines.forEach(function (lineNumber) {
+		editor.removeLineClass(lineNumber - 3, "background", "bg-coverage-warning");
+		editor.removeLineClass(lineNumber - 3, "background", "bg-coverage-success");
+		editor.addLineClass	(lineNumber - 3, "background", "bg-coverage-danger");
 	});
 
-	partiallyCoveredLines.forEach(function (lineNumber) { 
-		editor.removeLineClass(lineNumber - 2, "gutter", "bg-danger");
-		editor.removeLineClass(lineNumber - 2, "gutter", "bg-success");
-		editor.addLineClass	(lineNumber - 2, "gutter", "bg-warning");
+	partiallyCoveredLines.forEach(function (lineNumber) {
+		editor.removeLineClass(lineNumber - 3, "background", "bg-coverage-danger");
+		editor.removeLineClass(lineNumber - 3, "background", "bg-coverage-success");
+		editor.addLineClass	(lineNumber - 3, "background", "bg-coverage-warning");
 	});
+
+	// Forzo il refresh dell'editor, altrimenti il background viene caricato solo dopo aver scollato
+	editor.refresh();
 }
 
-// Funzione per ottenere i dati del form da inviare
-function getFormData() {
-	const formData = new FormData();
-	formData.append("testingClassCode", editor_utente.getValue());
-	formData.append("playerId", String(parseJwt(getCookie("jwt")).userId));
-	formData.append("mode", "sfida"); //Gestire modalità di gioco 
-	return formData;
-}
-
-async function ajaxRequest(
+async function ajaxRequest_ForRun(
 	url,
 	method = "POST",
 	data = null,
-	isJson = true,
+	isFormData = false,
 	dataType = "json"
 ) {
 	try {
 		const options = {
 			url: url,
 			type: method,
-			dataType: dataType,
-			processData: isJson, // Set to true to encode data properly
-			contentType: isJson
-				? "application/x-www-form-urlencoded; charset=UTF-8"
-				: false,
-			data: isJson && data ? $.param(data) : data, // Convert data to URL-encoded string
+			processData: false, // Se è FormData, non elaborarlo
+			contentType: "application/json",
+			data: JSON.stringify(data), // Passa i dati direttamente senza modificarli
 		};
 
 		const response = await $.ajax(options);
@@ -354,6 +391,34 @@ async function ajaxRequest(
 		throw error;
 	}
 }
+
+/*
+async function ajaxRequest_ForRun(
+	url,
+	method = "POST",
+	data = null,
+	isFormData = false,
+	dataType = "json"
+) {
+	try {
+		const options = {
+			url: url,
+			type: method,
+			processData: false, // Se è FormData, non elaborarlo
+			contentType: false,
+			data: data, // Passa i dati direttamente senza modificarli
+		};
+
+		const response = await $.ajax(options);
+		return response;
+	} catch (error) {
+		console.error("Si è verificato un errore:", error);
+		throw error;
+	}
+}
+
+ */
+
 
 function controlloScalata(
 	iswin,
@@ -483,66 +548,105 @@ function controlloScalata(
 			});
 	}
 }
+// === FUNZIONI PER LA GESTIONE DEL TIMER DELLA PARTITA SINGOLA ===
 
-// Funzione per analizzare l'output di Maven
-function parseMavenOutput(output) {
-    const lines = output.split('\n');
-    let results = {
-        errors: 0,
-        warnings: 0
-    };
+window.addEventListener("load", function () {
+	const mode = GetMode();
+	if (mode === "PartitaSingola")
+		startTimer();
+});
 
-    lines.forEach(line => {
-        // Regex per contare avvisi
-        const warningMatch = line.match(/^\[INFO\] (\d+) warning/);
-        if (warningMatch) {
-            results.warnings = parseInt(warningMatch[1], 10);
-        }
+function startTimer() {
+	console.log("StartTimer: ", timer_remainingTime);
+	timer = setInterval(function () {
+		document.getElementById('timerDisplay').innerHTML = formatTime(timer_remainingTime);
 
-        // Regex per contare errori
-        const errorMatch = line.match(/^\[INFO\] (\d+) error/);
-        if (errorMatch) {
-            results.errors = parseInt(errorMatch[1], 10);
-        }
-    });
-
-	document.getElementById("error_compiler").textContent = results.errors;
-	document.getElementById("warning_compiler").textContent =  results.warnings;
-    return results;
+		if (timer_remainingTime <= 0) {
+			clearInterval(timer);
+			onTimerEnd();
+		} else {
+			timer_remainingTime--;
+		}
+	}, 1000);
 }
 
-function pulisciLocalStorage(chiave) {
-    // Controlla se la chiave esiste nel localStorage
-    if (localStorage.getItem(chiave)) {
-        // Rimuovi la chiave dal localStorage
-        localStorage.removeItem(chiave);
-        console.log(`Dati associati a "${chiave}" rimossi dal localStorage.`);
-    } else {
-        console.log(`Nessun dato trovato per la chiave "${chiave}".`);
-    }
+function stopTimer() {
+	if (timer !== null) {
+		clearInterval(timer);
+		timer = null;
+	}
 }
 
-//Funzione per fare il replace del testo dell'editor 
+function formatTime(seconds) {
+	let hrs = Math.floor(seconds / 3600);
+	let mins = Math.floor((seconds % 3600) / 60);
+	let secs = seconds % 60;
+	return (
+		String(hrs).padStart(2, '0') + ":" +
+		String(mins).padStart(2, '0') + ":" +
+		String(secs).padStart(2, '0')
+	);
+}
+
+function onTimerEnd() {
+	isActionInProgress = true; // Imposta il flag per bloccare altre azioni
+	run_button.disabled = true; // Disabilita il pulsante di esecuzione
+	coverage_button.disabled = true; // Disabilita il pulsante di coverage
+	openModalWithText(
+		"Oh no! Il tempo è scaduto!",
+		"Vuoi consegnare il codice corrente o mantenere l'ultimo compilato?",
+		[
+			{ tagName: "button", text: "Consegna", class: 'btn btn-primary', data_bs_dismiss: "modal", onclick: () => handleGameAction(true, true) },
+			{ tagName: "button", text: "Mantieni", class: 'btn btn-primary', data_bs_dismiss: "modal", onclick: () => handleGameAction(true, false) }
+		]
+	);
+}
+
+
+// === FUNZIONI DI UTILITÀ PER L'EDITOR E LA SESSIONE ===
+
+// Inizializzo il contenuto dell'editor con il codice salvato nella sessione
+document.addEventListener('DOMContentLoaded', () => {
+	console.log("sessione: ", previousGameObject);
+	if (previousGameObject && previousGameObject.testingClassCode) {
+		editor_utente.setValue(previousGameObject.testingClassCode);
+	}
+	if (previousGameObject && GetMode() === "PartitaSingola" && previousGameObject.remainingTime) {
+		timer_remainingTime = previousGameObject.remainingTime;
+	}
+});
+
+//Funzione per fare il replace del testo dell'editor
 function replaceText(text, replacements) {
     return text.replace(/\b(TestClasse|username|userID|date)\b/g, match => replacements[match] || match);
 }
 
-//Funzione carica
-function SetInitialEditor(replacements){
-	let text = editor_utente.getValue();
-	let newContent = replaceText(text, replacements) ;
-	editor_utente.setValue(newContent);
+function SetInitialEditor(replacements) {
+    const text = editor_utente.getValue();
+    console.log("[SetInitialEditor] Testo originale:", text);
+    const newContent = replaceText(text, replacements);
+    console.log("[SetInitialEditor] Testo aggiornato:", newContent);
+    editor_utente.setValue(newContent);
 }
 
 //Ottieni parametro dal URL
 function getParameterByName(name) {
-	const url = window.location.href;
-	name = name.replace(/[\[\]]/g, "\\$&");
-	const regex = new RegExp("[?&]" + name + "(=([^&#]*)|&|#|$)");
-	const results = regex.exec(url);
-	if (!results) return null;
-	if (!results[2]) return "";
-	return decodeURIComponent(results[2].replace(/\+/g, " "));
+    const url = window.location.href;
+    name = name.replace(/[\[\]]/g, "\\$&");
+    const regex = new RegExp("[?&]" + name + "(=([^&#]*)|&|#|$)");
+    const results = regex.exec(url);
+    if (!results) return null;
+    if (!results[2]) return "";
+    return decodeURIComponent(results[2].replace(/\+/g, " "));
+}
+
+function GetMode() {
+    const mode = getParameterByName("mode");
+    if (mode) {
+        const cleanedMode = mode.replace(/[^\w]/g, "").trim(); // Rimuove caratteri non alfanumerici
+        return cleanedMode;
+    }
+    return null;
 }
 
 // modal info
@@ -558,11 +662,22 @@ function openModalWithText(text_title, text_content, buttons = []) {
 	// Aggiungi bottoni personalizzati se sono stati forniti
 	if (buttons.length > 0) {
 		buttons.forEach(button => {
-			let btn = document.createElement('a');
+			let btn = document.createElement(button.tagName);
 			btn.innerText = button.text;
-			btn.href = button.href;  // Assegna il link al pulsante
+
 			btn.className = button.class || 'btn btn-primary'; // Classe di default se non specificata
 			btn.target = button.target || '_self'; // Target opzionale, default è nella stessa finestra
+
+			if (button.data_bs_dismiss)
+				btn.setAttribute('data-bs-dismiss', button.data_bs_dismiss); // Aggiungo il pulsante di chiusura del modale oltre alla "X"
+
+			if (button.href)
+				btn.href = button.href;  // Assegna il link al pulsante
+
+			if (button.onclick) {
+				btn.addEventListener("click", button.onclick); // Assegna la funzione all'evento
+			}
+
 			modalFooter.appendChild(btn);
 		});
 	}
@@ -601,4 +716,19 @@ function openModalError(text_title, text_content, buttons = []) {
 
 	// Mostra il modal
 	modal.show();
+}
+
+function objectToFormData(obj, formData = new FormData(), parentKey = '') {
+	for (let key in obj) {
+		if (obj.hasOwnProperty(key)) {
+			let fullKey = parentKey ? `${parentKey}[${key}]` : key;
+
+			if (typeof obj[key] === 'object' && obj[key] !== null && !(obj[key] instanceof File)) {
+				objectToFormData(obj[key], formData, fullKey);
+			} else {
+				formData.append(fullKey, obj[key]);
+			}
+		}
+	}
+	return formData;
 }
